third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-element-distributed-nodes-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests for content element's getDistributedNodes().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
